# docs/swagger.yaml
basePath: /
definitions:
...
  model.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  model.WebhookRequest:
    properties:
      events:
        items:
          type: string
        type: array
      webhookurl:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@fiozap.io
    name: FioZap Support
  description: WhatsApp API using whatsmeow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FioZap API
  version: "1.0"
paths:
...
  /chat/presence:
    post:
      consumes:
      - application/json
      description: Send typing/recording indicator to a chat
      parameters:
      - description: Chat presence data
        in: body
        name: request
        required: true
        schema:
          properties:
            media:
              type: string
            phone:
              type: string
            state:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send chat presence
      tags:
      - User
...
  /sessions/{sessionId}/webhook:
    delete:
      description: Remove webhook URL and unsubscribe from all events for the session
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete webhook configuration
      tags:
      - Webhook
    get:
      description: Get current webhook URL and subscribed events for the session
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get webhook configuration
      tags:
      - Webhook
    post:
      consumes:
      - application/json
      description: Configure webhook URL and events to subscribe for the session
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      - description: Webhook configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.WebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Set webhook configuration
      tags:
      - Webhook
    put:
      consumes:
      - application/json
      description: Update webhook URL, events and active status for the session
      parameters:
      - description: Session name
        in: path
        name: sessionId
        required: true
        type: string
      - description: Webhook update data
        in: body
        name: request
        required: true
        schema:
          properties:
            active:
              type: boolean
            events:
              items:
                type: string
              type: array
            webhook:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Update webhook configuration
      tags:
      - Webhook
  /user/avatar:
    post:
      consumes:
      - application/json
      description: Get profile picture URL for a phone number
      parameters:
      - description: Phone and preview option
        in: body
        name: request
        required: true
        schema:
          properties:
            phone:
              type: string
            preview:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user avatar
      tags:
      - User
  /user/check:
    post:
      consumes:
      - application/json
      description: Check if phone numbers are registered on WhatsApp
      parameters:
      - description: Phone numbers
        in: body
        name: request
        required: true
        schema:
          properties:
            phone:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Check if users are on WhatsApp
      tags:
      - User
  /user/contacts:
    get:
      description: Get all contacts from the connected WhatsApp account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get contacts
      tags:
      - User
  /user/info:
    post:
      consumes:
      - application/json
      description: Get WhatsApp user info by phone numbers
      parameters:
      - description: Phone numbers
        in: body
        name: request
        required: true
        schema:
          properties:
            phone:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Get user info
      tags:
      - User
  /user/presence:
    post:
      consumes:
      - application/json
      description: Send online/offline presence status
      parameters:
      - description: Presence (available/unavailable)
        in: body
        name: request
        required: true
        schema:
          properties:
            presence:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
      security:
      - ApiKeyAuth: []
      summary: Send presence
      tags:
      - User
securityDefinitions:
  AdminKeyAuth:
    description: Admin token for user management
    in: header
    name: Authorization
    type: apiKey
  ApiKeyAuth:
    description: User API token for session operations
    in: header
    name: Token
    type: apiKey
swagger: "2.0"